requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.11
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.1
idna==3.10
kombu==5.4.2
multidict==6.1.0
prompt_toolkit==3.0.50
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyvips==2.2.3
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.31.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.17.1
